refactor: LayersMenu refactoring done
