+++ src/components/LayersMenu/LayersMenu.tsx
@@ -1,13 +1,25 @@
 import { LayersList } from './LayersList';
 import { OpacitySlider } from './OpacitySlider';
-import { AddLayerButton } from './AddLayerButton';
+import { useAppDispatch } from '@/redux/hooks';
+import { PlusOutlined } from '@ant-design/icons';
+import { addLayer } from '@/redux/layers';
+import { Button } from 'antd';
 
 export function LayersMenu() {
+  const d = useAppDispatch();
+
   return (
     <div className='h-1/2'>
       <div className='h-full flex flex-col'>
         <h1 className='m-2 text-2xl text-center'>Слои</h1>
-        <AddLayerButton />
+        {/* Add layer button */}
+        <div className='m-2'>
+          <Button
+            icon={<PlusOutlined />}
+            block
+            onClick={() => d(addLayer())}></Button>
+        </div>
+
         <OpacitySlider />
         <LayersList />
       </div>
